Collapse service wrappers and split the settings dialog
unwrapCommandResult in daylite-service-helpers.ts replaces the
per-service error readers; every service wrapper is now a one-liner
with its German fallback message unchanged.
display-settings.ts shrinks to loadDisplaySettings and
saveDisplaySettings(patch), which also removes the duplicated
default-value comments; the spec covers the same merge behavior.
settings-dialog.tsx is split into components/settings/ with one file
per panel and a shared StatusAlert.
