+++ src/services/display-settings.spec.ts
@@ -1,10 +1,5 @@
 import { beforeEach, describe, expect, it, mock } from "bun:test";
-import {
-  loadHideNonPlannableEmployees,
-  loadShowWeekend,
-  saveHideNonPlannableEmployees,
-  saveShowWeekend,
-} from "./display-settings";
+import { loadDisplaySettings, saveDisplaySettings } from "./display-settings";
 
 interface DisplaySettings {
   hideNonPlannableEmployees?: boolean;
@@ -37,38 +32,38 @@ describe("display settings service", () => {
     mockSaveLocalStore.mockClear();
   });
 
-  it("saveHideNonPlannableEmployees preserves an already-stored showWeekend value", async () => {
+  it("saving one setting preserves the other stored settings", async () => {
     store.displaySettings = {
       hideNonPlannableEmployees: true,
       showWeekend: true,
     };
 
-    await saveHideNonPlannableEmployees(false);
+    await saveDisplaySettings({ hideNonPlannableEmployees: false });
 
     expect(store.displaySettings?.showWeekend).toBe(true);
     expect(store.displaySettings?.hideNonPlannableEmployees).toBe(false);
   });
 
-  it("loadHideNonPlannableEmployees defaults to true when unset", async () => {
-    expect(await loadHideNonPlannableEmployees()).toBe(true);
+  it("loads the backend defaults when nothing is stored", async () => {
+    const settings = await loadDisplaySettings();
+
+    expect(settings.hideNonPlannableEmployees).toBe(true);
+    expect(settings.showWeekend).toBe(false);
   });
 
-  it("loadShowWeekend defaults to false when unset", async () => {
-    expect(await loadShowWeekend()).toBe(false);
+  it("persists a saved value", async () => {
+    await saveDisplaySettings({ showWeekend: true });
+
+    expect((await loadDisplaySettings()).showWeekend).toBe(true);
   });
 
-  it("saveShowWeekend persists the value", async () => {
-    await saveShowWeekend(true);
-    expect(await loadShowWeekend()).toBe(true);
-  });
-
-  it("saveShowWeekend preserves an already-stored hideNonPlannableEmployees value", async () => {
+  it("saving showWeekend preserves an already-stored hideNonPlannableEmployees value", async () => {
     store.displaySettings = {
       hideNonPlannableEmployees: false,
       showWeekend: false,
     };
 
-    await saveShowWeekend(true);
+    await saveDisplaySettings({ showWeekend: true });
 
     expect(store.displaySettings?.showWeekend).toBe(true);
     expect(store.displaySettings?.hideNonPlannableEmployees).toBe(false);
