+++ src/services/display-settings.ts
@@ -1,77 +1,39 @@
 import { commands, type DisplaySettings } from "../generated/tauri";
+import { unwrapCommandResult } from "./daylite-service-helpers";
 
-// Mirrors the Rust default (DisplaySettings::default): hide non-plannable
-// employees unless the user has explicitly turned the toggle off.
-const DEFAULT_HIDE_NON_PLANNABLE_EMPLOYEES = true;
-
-// Mirrors the Rust default (DisplaySettings::default): the weekend is hidden
-// unless the user has explicitly turned the toggle on.
-const DEFAULT_SHOW_WEEKEND = false;
-
-const DEFAULT_DISPLAY_SETTINGS: DisplaySettings = {
-  hideNonPlannableEmployees: DEFAULT_HIDE_NON_PLANNABLE_EMPLOYEES,
-  showWeekend: DEFAULT_SHOW_WEEKEND,
+// Mirrors DisplaySettings::default() in the Rust backend.
+const defaultDisplaySettings: DisplaySettings = {
+  hideNonPlannableEmployees: true,
+  showWeekend: false,
 };
 
-export async function loadHideNonPlannableEmployees(): Promise<boolean> {
-  const result = await commands.loadLocalStore();
-  if (result.status === "error") {
-    throw new Error(result.error.userMessage);
-  }
+const loadErrorMessage =
+  "Die lokale Konfiguration konnte nicht geladen werden.";
 
-  return (
-    result.data.displaySettings?.hideNonPlannableEmployees ??
-    DEFAULT_HIDE_NON_PLANNABLE_EMPLOYEES
+export async function loadDisplaySettings(): Promise<DisplaySettings> {
+  const store = unwrapCommandResult(
+    await commands.loadLocalStore(),
+    loadErrorMessage,
   );
+  return { ...defaultDisplaySettings, ...store.displaySettings };
 }
 
-export async function saveHideNonPlannableEmployees(
-  hideNonPlannableEmployees: boolean,
+export async function saveDisplaySettings(
+  patch: Partial<DisplaySettings>,
 ): Promise<void> {
-  const loaded = await commands.loadLocalStore();
-  if (loaded.status === "error") {
-    throw new Error(loaded.error.userMessage);
-  }
-
-  const saved = await commands.saveLocalStore({
-    ...loaded.data,
-    // Merge into the existing display settings so saving one field does not drop
-    // the others (e.g. showWeekend).
-    displaySettings: {
-      ...(loaded.data.displaySettings ?? DEFAULT_DISPLAY_SETTINGS),
-      hideNonPlannableEmployees,
-    },
-  });
-  if (saved.status === "error") {
-    throw new Error(saved.error.userMessage);
-  }
-}
-
-export async function loadShowWeekend(): Promise<boolean> {
-  const result = await commands.loadLocalStore();
-  if (result.status === "error") {
-    throw new Error(result.error.userMessage);
-  }
-
-  return result.data.displaySettings?.showWeekend ?? DEFAULT_SHOW_WEEKEND;
-}
-
-export async function saveShowWeekend(showWeekend: boolean): Promise<void> {
-  const loaded = await commands.loadLocalStore();
-  if (loaded.status === "error") {
-    throw new Error(loaded.error.userMessage);
-  }
-
-  const saved = await commands.saveLocalStore({
-    ...loaded.data,
-    // Merge into the existing display settings so saving one field does not drop
-    // the others (e.g. hideNonPlannableEmployees).
-    displaySettings: {
-      ...(loaded.data.displaySettings ?? DEFAULT_DISPLAY_SETTINGS),
-      showWeekend,
-    },
-  });
-  if (saved.status === "error") {
-    throw new Error(saved.error.userMessage);
-  }
+  const store = unwrapCommandResult(
+    await commands.loadLocalStore(),
+    loadErrorMessage,
+  );
+  unwrapCommandResult(
+    await commands.saveLocalStore({
+      ...store,
+      displaySettings: {
+        ...defaultDisplaySettings,
+        ...store.displaySettings,
+        ...patch,
+      },
+    }),
+    "Die Anzeige-Einstellung konnte nicht gespeichert werden.",
+  );
 }
